Fixing lint warnings.
